[kbn-expandable-flyout] - refactor internal redux store to accommodate new ui changes coming soon
