+++ packages/kbn-expandable-flyout/src/hooks/use_expandable_flyout_api.ts
@@ -21,8 +21,8 @@ import {
   openPreviewPanelAction,
   openRightPanelAction,
   previousPreviewPanelAction,
-} from '../actions';
-import { useDispatch } from '../redux';
+} from '../store/actions';
+import { useDispatch } from '../store/redux';
 import { FlyoutPanelProps, type ExpandableFlyoutApi } from '../types';
 
 export type { ExpandableFlyoutApi };
