+++ packages/kbn-expandable-flyout/src/provider.test.tsx
@@ -11,8 +11,8 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { TestProvider } from './test/provider';
 import { UrlSynchronizer } from './provider';
-import * as actions from './actions';
-import { State } from './state';
+import * as actions from './store/actions';
+import { State } from './store/state';
 import { of } from 'rxjs';
 
 const mockGet = jest.fn();
@@ -28,14 +28,18 @@ describe('UrlSynchronizer', () => {
     const urlChangedAction = jest.spyOn(actions, 'urlChangedAction');
 
     const initialState: State = {
-      byId: {
-        [urlKey]: {
-          right: { id: 'key1' },
-          left: { id: 'key11' },
-          preview: undefined,
+      data: {
+        panels: {
+          byId: {
+            [urlKey]: {
+              right: { id: 'key1' },
+              left: { id: 'key11' },
+              preview: undefined,
+            },
+          },
+          needsSync: true,
         },
       },
-      needsSync: true,
     };
 
     render(
@@ -55,8 +59,12 @@ describe('UrlSynchronizer', () => {
       change$: mockChange$,
     });
     const initialState: State = {
-      byId: {},
-      needsSync: true,
+      data: {
+        panels: {
+          byId: {},
+          needsSync: true,
+        },
+      },
     };
 
     render(
@@ -81,14 +89,18 @@ describe('UrlSynchronizer', () => {
       change$: mockChange$,
     });
     const initialState: State = {
-      byId: {
-        [urlKey]: {
-          right: { id: 'key1' },
-          left: { id: 'key2' },
-          preview: undefined,
+      data: {
+        panels: {
+          byId: {
+            [urlKey]: {
+              right: { id: 'key1' },
+              left: { id: 'key2' },
+              preview: undefined,
+            },
+          },
+          needsSync: true,
         },
       },
-      needsSync: true,
     };
 
     render(
