+++ packages/kbn-expandable-flyout/src/store/reducers.ts
@@ -21,131 +21,131 @@ import {
   openPreviewPanelAction,
   urlChangedAction,
 } from './actions';
-import { initialState } from './state';
+import { initialDataState } from './state';
 
-export const reducer = createReducer(initialState, (builder) => {
+export const dataReducer = createReducer(initialDataState, (builder) => {
   builder.addCase(openPanelsAction, (state, { payload: { preview, left, right, id } }) => {
-    if (id in state.byId) {
-      state.byId[id].right = right;
-      state.byId[id].left = left;
-      state.byId[id].preview = preview ? [preview] : undefined;
+    if (id in state.panels.byId) {
+      state.panels.byId[id].right = right;
+      state.panels.byId[id].left = left;
+      state.panels.byId[id].preview = preview ? [preview] : undefined;
     } else {
-      state.byId[id] = {
+      state.panels.byId[id] = {
         left,
         right,
         preview: preview ? [preview] : undefined,
       };
     }
 
-    state.needsSync = true;
+    state.panels.needsSync = true;
   });
 
   builder.addCase(openLeftPanelAction, (state, { payload: { left, id } }) => {
-    if (id in state.byId) {
-      state.byId[id].left = left;
+    if (id in state.panels.byId) {
+      state.panels.byId[id].left = left;
     } else {
-      state.byId[id] = {
+      state.panels.byId[id] = {
         left,
         right: undefined,
         preview: undefined,
       };
     }
 
-    state.needsSync = true;
+    state.panels.needsSync = true;
   });
 
   builder.addCase(openRightPanelAction, (state, { payload: { right, id } }) => {
-    if (id in state.byId) {
-      state.byId[id].right = right;
+    if (id in state.panels.byId) {
+      state.panels.byId[id].right = right;
     } else {
-      state.byId[id] = {
+      state.panels.byId[id] = {
         right,
         left: undefined,
         preview: undefined,
       };
     }
 
-    state.needsSync = true;
+    state.panels.needsSync = true;
   });
 
   builder.addCase(openPreviewPanelAction, (state, { payload: { preview, id } }) => {
-    if (id in state.byId) {
-      if (state.byId[id].preview) {
-        const previewIdenticalToLastOne = deepEqual(preview, state.byId[id].preview?.at(-1));
-        // Only append preview when it does not match the last item in state.byId[id].preview
+    if (id in state.panels.byId) {
+      if (state.panels.byId[id].preview) {
+        const previewIdenticalToLastOne = deepEqual(preview, state.panels.byId[id].preview?.at(-1));
+        // Only append preview when it does not match the last item in state.data.byId[id].preview
         if (!previewIdenticalToLastOne) {
-          state.byId[id].preview?.push(preview);
+          state.panels.byId[id].preview?.push(preview);
         }
       } else {
-        state.byId[id].preview = preview ? [preview] : undefined;
+        state.panels.byId[id].preview = preview ? [preview] : undefined;
       }
     } else {
-      state.byId[id] = {
+      state.panels.byId[id] = {
         right: undefined,
         left: undefined,
         preview: preview ? [preview] : undefined,
       };
     }
 
-    state.needsSync = true;
+    state.panels.needsSync = true;
   });
 
   builder.addCase(previousPreviewPanelAction, (state, { payload: { id } }) => {
-    if (id in state.byId) {
-      state.byId[id].preview?.pop();
+    if (id in state.panels.byId) {
+      state.panels.byId[id].preview?.pop();
     }
 
     // if state is stored in url, click go back in preview should utilize browser history
-    state.needsSync = false;
+    state.panels.needsSync = false;
   });
 
   builder.addCase(closePanelsAction, (state, { payload: { id } }) => {
-    if (id in state.byId) {
-      state.byId[id].right = undefined;
-      state.byId[id].left = undefined;
-      state.byId[id].preview = undefined;
+    if (id in state.panels.byId) {
+      state.panels.byId[id].right = undefined;
+      state.panels.byId[id].left = undefined;
+      state.panels.byId[id].preview = undefined;
     }
 
-    state.needsSync = true;
+    state.panels.needsSync = true;
   });
 
   builder.addCase(closeLeftPanelAction, (state, { payload: { id } }) => {
-    if (id in state.byId) {
-      state.byId[id].left = undefined;
+    if (id in state.panels.byId) {
+      state.panels.byId[id].left = undefined;
     }
 
-    state.needsSync = true;
+    state.panels.needsSync = true;
   });
 
   builder.addCase(closeRightPanelAction, (state, { payload: { id } }) => {
-    if (id in state.byId) {
-      state.byId[id].right = undefined;
+    if (id in state.panels.byId) {
+      state.panels.byId[id].right = undefined;
     }
 
-    state.needsSync = true;
+    state.panels.needsSync = true;
   });
 
   builder.addCase(closePreviewPanelAction, (state, { payload: { id } }) => {
-    if (id in state.byId) {
-      state.byId[id].preview = undefined;
+    if (id in state.panels.byId) {
+      state.panels.byId[id].preview = undefined;
     }
 
-    state.needsSync = true;
+    state.panels.needsSync = true;
   });
 
   builder.addCase(urlChangedAction, (state, { payload: { preview, left, right, id } }) => {
-    if (id in state.byId) {
-      state.byId[id].right = right;
-      state.byId[id].left = left;
-      state.byId[id].preview = preview ? [preview] : undefined;
+    if (id in state.panels.byId) {
+      state.panels.byId[id].right = right;
+      state.panels.byId[id].left = left;
+      state.panels.byId[id].preview = preview ? [preview] : undefined;
     } else {
-      state.byId[id] = {
+      state.panels.byId[id] = {
         right,
         left,
         preview: preview ? [preview] : undefined,
       };
     }
 
-    state.needsSync = false;
+    state.panels.needsSync = false;
   });
 });
