+++ packages/kbn-expandable-flyout/src/store/redux.ts
@@ -11,13 +11,14 @@ import { createContext } from 'react';
 import { createDispatchHook, createSelectorHook, ReactReduxContextValue } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
 import { createSelector } from 'reselect';
-import { reducer } from './reducer';
+import { dataReducer } from './reducers';
 import { initialState, State } from './state';
 
 export const store = configureStore({
-  reducer,
+  reducer: {
+    data: dataReducer,
+  },
   devTools: process.env.NODE_ENV !== 'production',
-  enhancers: [],
 });
 
 export const Context = createContext<ReactReduxContextValue<State>>({
@@ -30,7 +31,9 @@ export const useSelector = createSelectorHook(Context);
 
 const stateSelector = (state: State) => state;
 
-export const selectPanelsById = (id: string) =>
-  createSelector(stateSelector, (state) => state.byId[id] || {});
+const dataSelector = createSelector(stateSelector, (state) => state.data);
 
-export const selectNeedsSync = () => createSelector(stateSelector, (state) => state.needsSync);
+const panelsSelector = createSelector(dataSelector, (state) => state.panels);
+export const selectPanelsById = (id: string) =>
+  createSelector(panelsSelector, (state) => state.byId[id] || {});
+export const selectNeedsSync = () => createSelector(panelsSelector, (state) => state.needsSync);
